Fix bug that prevented writes from being noticed
# What

The command execution would never happen for files that had their contents change. This is because I only tested this script with `touch` which just changed the modification time.

# Why

The source of the bug was `file = File.new(options[:file])`. Assigning the file to a variable cached all of its attributes in the object. So when I passed file to the old_mtime and new_mtime checks it would return the same mtime because it was never updated in between the sleep interval.

The easy fix was just to open the file again after the wait.

# TODO

Move the logic for retrieving the mtime to a function that handles opening the file, getting the mtime, closing the file and returning the mtime.